requirements.txt
Django==1.4
South==0.7.5
boto==2.5.2
distribute==0.6.19
dj-database-url==0.2.0
django-crispy-forms==1.1.4
django-storages==1.1.4
gunicorn==0.14.5
psycopg2==2.4.5
wsgiref==0.1.2
